Book Store exercise
This one was a little bit more complicated than the other ones. I wasn't able to do it by myself so I looked up for community's solutions and ended up trying to understand the solution as well as the algorithms needed to do the exercise in a proper way. I found out how 'use' and 'assert' keyword works in a given scenarios. I've commented the code in a way I understand how those things works.